+++ src/cli/swpm/commands/add.js
@@ -1,4 +1,4 @@
-import { translateArgs } from '../../helpers/args.js'
+import { translateArgs } from '../../../helpers/args.js'
 
 const middleware = (yargs) => {
   if ('save-dev' in yargs) {
@@ -12,13 +12,17 @@ const middleware = (yargs) => {
   if ('save-peer' in yargs) {
     translateArgs(yargs, '--save-peer')
   }
+
+  if ('save-exact' in yargs) {
+    translateArgs(yargs, '--save-exact', '-E')
+  }
 }
 
-const remove = {
-  command: 'remove <package> [args] [FLAGS]',
-  aliases: ['r', 'rm', 'uninstall', 'un'],
-  desc: 'remove package',
-  conflicts: ['add', 'clean', 'install', 'update', 'upgrade'],
+const add = {
+  command: 'add <package> [args] [FLAGS]',
+  aliases: ['a'],
+  desc: 'add package',
+  conflicts: ['clean', 'install', 'remove', 'update', 'upgrade'],
   builder: (yargs) => {
     yargs.positional('package', {
       type: 'string',
@@ -28,8 +32,8 @@ const remove = {
     yargs.option('save-dev', {
       alias: 'D',
       type: 'boolean',
-      desc: 'remove package as devDependencies',
-      usage: '$0 remove <package> --save-dev',
+      desc: 'add package as devDependencies',
+      usage: '$0 add <package> --save-dev',
       implies: ['package'],
       conflicts: ['save-optional', 'save-peer']
     })
@@ -37,25 +41,33 @@ const remove = {
     yargs.option('save-optional', {
       alias: 'O',
       type: 'boolean',
-      desc: 'remove package as optionalDependencies',
-      usage: '$0 remove <package> --save-optional',
+      desc: 'add package as optionalDependencies',
+      usage: '$0 add <package> --save-optional',
       implies: ['package'],
       conflicts: ['save-dev', 'save-peer']
     })
 
     yargs.option('save-peer', {
       type: 'boolean',
-      desc: 'remove package as peerDependencies',
-      usage: '$0 remove <package> --save-peer',
+      desc: 'add package as peerDependencies',
+      usage: '$0 add <package> --save-peer',
       implies: ['package'],
       conflicts: ['save-dev', 'save-optional']
+    })
+
+    yargs.option('save-exact', {
+      alias: 'E',
+      type: 'boolean',
+      desc: 'add package as exact version',
+      usage: '$0 add <package> --save-exact',
+      implies: ['package']
     })
 
     yargs.option('global', {
       alias: 'g',
       type: 'boolean',
-      desc: 'remove package as global',
-      usage: '$0 remove <package> --global',
+      desc: 'add package as global',
+      usage: '$0 add <package> --global',
       implies: ['package']
     })
 
@@ -65,4 +77,4 @@ const remove = {
   }
 }
 
-export default remove
+export default add