+++ src/cli/swpm/config.js
@@ -1,19 +1,21 @@
+import path from 'node:path';
+
 import { argv } from 'node:process'
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
 
 import { commands, options } from './cli.js'
-import middleware from './middleware.js'
+import middleware from '../middleware.js'
 
 const config = await yargs(hideBin(argv))
-  .scriptName("swpm")
+  .scriptName(path.basename(argv[1], path.extname(argv[1])))
   .commands(commands)
   .options(options)
   .middleware(middleware)
   .usage('$0 [<command>] [--] [args] [FLAGS]')
   .help()
   .version(false)
-  .epilog('dein Software - copyright 2022')
+  .epilog('dein Software © 2022')
   .argv
 
 export default config