+++ src/cli/swpx/config.js
@@ -1,0 +1,20 @@
+import path from 'node:path';
+
+import { argv } from 'node:process'
+import yargs from 'yargs'
+import { hideBin } from 'yargs/helpers'
+
+import { options } from './cli.js'
+import middleware from '../middleware.js'
+
+const config = await yargs(hideBin(argv))
+  .scriptName(path.basename(argv[1], path.extname(argv[1])))
+  .options(options)
+  .middleware(middleware)
+  .usage('$0 [<command>] [FLAGS]')
+  .help()
+  .version(false)
+  .epilog('dein Software © 2022')
+  .argv
+
+export default config