+++ src/flags/test.js
@@ -1,7 +1,7 @@
 import { exit } from 'node:process'
 import { showCommand } from '../helpers/cmds.js'
 
-export const testCommand = (pkg) => {
-  showCommand(pkg)
+export const testCommand = ($0, pkg) => {
+  showCommand($0, pkg)
   exit(0)
 }